+++ app/src/main/java/org/twinlife/twinme/ui/conversationFilesActivity/FullscreenVideoViewHolder.java
@@ -86,7 +86,7 @@ public class FullscreenVideoViewHolder extends RecyclerView.ViewHolder implement
                 float progressBarWidth = Design.DISPLAY_WIDTH - (DESIGN_PROGRESS_SIDE_MARGIN * Design.WIDTH_RATIO * 2f);
                 long interval = (long) (0.5f * duration / progressBarWidth);
                 mHandler.postDelayed(this, interval);
-                mSeekBar.setProgress((int) duration);
+                mSeekBar.setProgress(duration);
             }
         }
     };
