+++ app/src/main/java/org/twinlife/twinme/ui/settingsActivity/QualityOfServiceActivity.java
@@ -280,7 +280,7 @@ public class QualityOfServiceActivity extends AbstractOnboardingActivity {
         }
 
         float maxRecyclerViewHeight = Design.DISPLAY_HEIGHT - (MIN_CONTENT_VIEW_HEIGHT * Design.HEIGHT_RATIO) - lineHeight - Design.DOT_SIZE;
-        int textWidth = (int ) (Design.DISPLAY_WIDTH - (Design.ONBOARDING_TEXT_MARGIN * 2));
+        int textWidth = Design.DISPLAY_WIDTH - (Design.ONBOARDING_TEXT_MARGIN * 2);
 
         float recyclerViewHeight = 0;
         float minContentCellHeight = MIN_CONTENT_CELL_HEIGHT * Design.HEIGHT_RATIO;
